% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_CPAD.R
\name{analyse_CPAD}
\alias{analyse_CPAD}
\title{Analyse CPAD}
\usage{
analyse_CPAD(propab.matrix, propab.df, sumCases, avCases, saveplots = TRUE)
}
\arguments{
\item{propab.matrix:}{dataframe. The global active cases/country population (proprtional abundance) by country and day.}

\item{propab.df:}{dataframe. The proportional abundance dataframe of htree columns: \code{Country}, \code{Day}, and \code{propAb}.}

\item{sumCases:}{numeric vector. The daily sum of active SARS-CoV-2 cases; its length must be equal that the temporal slices present in \code{propab.df}.}

\item{avCases:}{numeric vector. The daily average of active SARS-CoV-2 cases; its length must be equal that the temporal slices present in \code{propab.df}.}

\item{saveplots:}{\code{TRUE} or \code{FALSE}. Save plots related to de distributions in your \code{wd}.Default \code{TRUE}.}
}
\value{
The function returns the \code{CPAD} dataframe to your Global Environment with the corresponding statistics for each day. If \code{saveplots} is \code{TRUE}, the \code{plots} folder is created in your \code{wd}, containing different folder with plots.
}
\description{
This function analyses the coronavirus proportional abundance distribution (CPAD)
}
\examples{
analyse_CPAD(propab.matrix=propab_matrix, propab.df=propab_df, sumCases=sumCases, avCases=avCases, saveplots= TRUE)



}
\author{
Simon P Castillo \email{spcastil@uc.cl}.
}
