% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseplot_shift.R
\name{analyseplot_shift}
\alias{analyseplot_shift}
\title{Estimate a single trajectory shift in the beta distribution estimates evolution}
\usage{
analyseplot_shift(
  CPAD,
  print.plot = TRUE,
  saveplot = TRUE,
  saveplot.ext = ".png"
)
}
\arguments{
\item{CPAD:}{dataframe. The dataframe returned by \code{\link{analyse_CPAD}}.}

\item{saveplot:}{\code{TRUE} or \code{FALSE}. Save the plots in your \code{wd}.Default \code{TRUE}.}

\item{saveplot.ext:}{character. The extension for the saved figures admitted by \code{\link[ggplot2]{ggsave}} (e.g., \code{".png"}, \code{".svg"}).Default \code{".png"}.}
}
\value{
This function returns to your global environment an object list named \code{plotbetatraj} with the plot(s) created. Also, if \code{saveplot} is \code{TRUE}, a folder named \code{betaplots} in the folder \code{plots} is created in your \code{wd}.

Also, this function returns a numeric object \code{plotShift} to your Global Environment, it is the esitmated date of the single transition.
}
\description{
This function calculates and plots a single shift in the temporal evolution of the estimated parameters of beta distribution $\hat{\alpha}$ and $\hat{\beta}$
}
\details{
The estimation of the single transition in the temporal trajectories of $\hat{\alpha}$ and $\hat{\beta}$ is performed via minimisaiton of the residual squared sum (RSS) (Bai. 1994).
}
\examples{
analyseplot_shift(CPAD, saveplot=TRUE, saveplot.ext= ".png")



}
\references{
J. M. Box-Steffensmeier, J. R. Freeman, M. P. Hitt, J. C. W. Pevehouse, Time Series Analysis for the Social Sciences (Cambridge University Press, 2014).

J. Bai, Least Squares Estimation of a Shift in Linear Processes. Journal of Time Series Analysis. 15, 453–472 (1994).
}
\author{
Simon P Castillo \email{spcastil@uc.cl}.
}
