% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_globalCOVID.R
\name{load_globalCOVID}
\alias{load_globalCOVID}
\title{Load global data}
\usage{
load_globalCOVID()
}
\value{
This function returns different objects to your Global Environment.

data_deaths: dataframe. The number of global deaths by day due to COVID.

data_positives: dataframe. The number of SARS-CoV-2(+).

data_recovered: dataframe. The number of recovered from SARS-CoV-2(+).

inc_matrix: dataframe. The global active cases by country and day.

propab_matrix:  dataframe The global active cases/country population by country and day.

inc_df: dataframe. A three column dataframe with country, date and the number of active cases.

propab_df: dataframe. A three column dataframe with country, date and the proportional abundance of active cases.

sumCases: numeric vector. The global sum of SARS-CoV-2(+) by day.

avCases: numeric vector. The global average of SARS-CoV-2(+) by day.



%ni%' <- Negate('%in%')
}
\description{
Load global data of SARS_CoV-2. This function obtain data from [CSSEGI-Johns Hopkins]{(https://github.com/CSSEGISandData/COVID-19} and it is built from other repos. [like this]{https://github.com/DrFabach/Corona}
}
\examples{
load_globalCOVID()
}
\author{
Simon P Castillo \email{spcastil@uc.cl}.
}
