% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratioCountries.R
\name{plot_ratioCountries}
\alias{plot_ratioCountries}
\title{Plot the fraction of countries with active SARS-CoV-2 vs sum #SARS-CoV-2, E|#SARS-CoV-2|, and E|prop ab|}
\usage{
plot_ratioCountries(CPAD, disease, saveplot = TRUE, saveplot.ext = ".png")
}
\arguments{
\item{CPAD:}{dataframe. The dataframe returned by \code{\link{analyse_CPAD}}.}

\item{disease:}{*character*. Name of the disease. It will apply if the plots are saved.}

\item{saveplots:}{\code{TRUE} or \code{FALSE}. Save the plots in your \code{wd}.Default \code{TRUE}.}

\item{saveplots.ext:}{character. The extension for the saved figures admitted by \code{\link[ggplot2]{ggsave}} (e.g., \code{".png"}, \code{".svg"}).Default \code{".png"}.}
}
\value{
This function return to your global environment an object list named \code{plotratioCountries} with the plot(s) created. Also, if \code{saveplot} is \code{TRUE}, a folder named \code{betaplots} in the folder \code{plots} is created in your \code{wd}.
}
\description{
This function plots the relations between the fraction of countries with active SARS-CoV-2 vs sum #SARS-CoV-2, E|#SARS-CoV-2|, and E|prop ab|.
}
\examples{
plot_ratioCountries(CPAD, saveplot=TRUE, saveplot.ext= ".png")



}
\author{
Simon P Castillo \email{spcastil@uc.cl}.
}
